requirements.txt
python-telegram-bot==20.0
psycopg2-binary==2.9.3
SQLAlchemy==2.0.18